Test legalEntity ainda não passa
git-svn-id: https://construtor.googlecode.com/svn/trunk@438 daf9df78-7b24-0410-bc62-633826f6319b

+++ park-jpa/core/src/main/java/park/model/Person.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public abstract class Person implements Serializable {
     private static final long serialVersionUID = 1L;
-    private long idPerson;
+    private int idPerson;
     private String address;
     private String name;
     private int version;
@@ -17,11 +17,11 @@ public abstract class Person implements Serializable {
         this.name = name;
     }
 
-    public long getIdPerson() {
+    public int getIdPerson() {
         return idPerson;
     }
 
-    public void setIdPerson(long idPerson) {
+    public void setIdPerson(int idPerson) {
         this.idPerson = idPerson;
     }
